Refactor ingesting script
Move functions into separate file
